LayoutTests/imported/w3c/web-platform-tests/navigation-api/state/cross-document-getState-expected.txt



FAIL entry.getState() still works for a non-current cross-document entry assert_equals: expected 2 but got 1
